JavaScript, PHP, HTML and Image Assets
Learn to implement two JSON client-servers with JavaScript, PHP and MySQL. This project includes explanation, source code and examples. The examples display information about animals in South Dakota, in response to one tap of a button, a menu selection, or you can see the entire table of animals. Learn two JSON techniques developers can use to send and retrieve data from a server. The first technique hard codes JSON values for return to the client. The second technique obtains JSON values from a MySQL table, to return to the client.